refactor Skill class
