+++ collated/functional/ChengSashankh.md
@@ -154,7 +154,7 @@ import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.job.Job;
 import seedu.address.model.job.JobTitle;
 import seedu.address.model.job.Location;
-import seedu.address.model.job.Skill;
+import seedu.address.model.skill.Skill;
 import seedu.address.model.tag.Tag;
 
 /***
