+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -22,7 +22,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.events.ui.RefreshReportPanelEvent;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.job.Skill;
+import seedu.address.model.skill.Skill;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Link;
