+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -15,7 +15,7 @@ import seedu.address.model.interview.InterviewLocation;
 import seedu.address.model.interview.InterviewTitle;
 import seedu.address.model.job.JobTitle;
 import seedu.address.model.job.Location;
-import seedu.address.model.job.Skill;
+import seedu.address.model.skill.Skill;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Link;
