+++ src/test/java/seedu/address/model/job/SkillsTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import seedu.address.model.skill.Skill;
 import seedu.address.testutil.Assert;
 
 public class SkillsTest {
