+++ src/test/java/seedu/address/storage/XmlAdaptedJobTest.java
@@ -12,7 +12,7 @@ import org.junit.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.job.JobTitle;
 import seedu.address.model.job.Location;
-import seedu.address.model.job.Skill;
+import seedu.address.model.skill.Skill;
 import seedu.address.testutil.Assert;
 
 public class XmlAdaptedJobTest {
