+++ src/test/java/seedu/address/testutil/EditJobDescriptorBuilder.java
@@ -8,7 +8,7 @@ import seedu.address.logic.commands.EditJobCommand;
 import seedu.address.model.job.Job;
 import seedu.address.model.job.JobTitle;
 import seedu.address.model.job.Location;
-import seedu.address.model.job.Skill;
+import seedu.address.model.skill.Skill;
 import seedu.address.model.tag.Tag;
 
 /**
